# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  mydb.StructNews:
    properties:
      admin_id:
        type: integer
      author:
        type: string
      class_id:
        type: integer
      comment_count:
        type: integer
      content:
        description: 文章内容
        type: string
      create_time:
        type: integer
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_headline:
        description: 是否头条
        type: boolean
      is_hot:
        description: 是否热门
        type: boolean
      is_recommended:
        description: 是否推荐
        type: boolean
      is_show:
        type: boolean
      keywords:
        type: string
      language:
        default: cn
        description: 文章使用的语言
        type: string
      sort:
        type: integer
      source:
        type: string
      status:
        type: integer
      subtitle:
        type: string
      title:
        type: string
      url:
        type: string
      view_count:
        type: integer
    type: object
...
  util.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: nav.fandoc.org
info:
  contact: {}
  description: This is a sample server for a nav web site.
  title: Nav Web Site API
  version: "1.0"
paths:
...
  /news/add:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 添加一条新的新闻记录
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 分类ID
        in: formData
        name: class_id
        required: true
        type: integer
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 副标题
        in: formData
        name: subtitle
        type: string
      - description: 链接
        in: formData
        name: url
        type: string
      - description: 描述
        in: formData
        name: description
        required: true
        type: string
      - description: 图标
        in: formData
        name: icon
        type: string
      - description: 关键词
        in: formData
        name: keywords
        type: string
      - description: 排序
        in: formData
        name: sort
        type: integer
      - description: 是否显示
        in: formData
        name: is_show
        type: boolean
      - description: 状态
        in: formData
        name: status
        type: integer
      - description: 作者
        in: formData
        name: author
        type: string
      - description: 来源
        in: formData
        name: source
        type: string
      - description: 语言
        in: formData
        name: language
        type: string
      - description: 是否热门
        in: formData
        name: is_hot
        type: boolean
      - description: 是否头条
        in: formData
        name: is_headline
        type: boolean
      - description: 是否推荐
        in: formData
        name: is_recommended
        type: boolean
      - description: 内容
        in: formData
        name: content
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 新闻添加成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: 添加新闻失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 添加新闻
      tags:
      - news
...
  /news/delete/{id}:
    delete:
      description: 根据新闻ID删除新闻
      parameters:
      - description: 新闻ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 新闻删除成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: 删除新闻失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 删除新闻
      tags:
      - news
...
  /news/detail/{id}:
    get:
      description: 根据新闻ID获取新闻详情
      parameters:
      - description: 新闻ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取新闻详情成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/mydb.StructNews'
                message:
                  type: string
              type: object
        "500":
          description: 获取新闻详情失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 获取新闻详情
      tags:
      - news
...
  /news/list:
    get:
      description: 获取新闻列表，支持按分类ID筛选
      parameters:
      - description: 新闻分类ID
        in: query
        name: class_id
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取新闻列表成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  items:
                    $ref: '#/definitions/mydb.StructNews'
                  type: array
                message:
                  type: string
              type: object
        "500":
          description: 获取新闻列表失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 获取新闻列表
      tags:
      - news
  /news/update/{id}:
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: 根据新闻ID编辑新闻内容
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 新闻ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类ID
        in: formData
        name: class_id
        required: true
        type: integer
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 副标题
        in: formData
        name: subtitle
        type: string
      - description: 链接地址
        in: formData
        name: url
        type: string
      - description: 描述
        in: formData
        name: description
        type: string
      - description: 图标
        in: formData
        name: icon
        type: string
      - description: 关键词
        in: formData
        name: keywords
        type: string
      - description: 排序
        in: formData
        name: sort
        type: integer
      - description: 是否显示
        in: formData
        name: is_show
        type: boolean
      - description: 状态
        in: formData
        name: status
        type: integer
      - description: 作者
        in: formData
        name: author
        type: string
      - description: 来源
        in: formData
        name: source
        type: string
      - description: 语言
        in: formData
        name: language
        type: string
      - description: 是否热门
        in: formData
        name: is_hot
        type: boolean
      - description: 是否头条
        in: formData
        name: is_headline
        type: boolean
      - description: 是否推荐
        in: formData
        name: is_recommended
        type: boolean
      - description: 内容
        in: formData
        name: content
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 新闻编辑成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: 编辑新闻失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 编辑新闻
      tags:
      - news
...
swagger: "2.0"